requirements.txt
django
djangorestframework
psycopg2
Faker
django-mptt